requirements.txt
BeautifulSoup==3.2.1
GnuPGInterface==0.3.2
IMAPClient==0.9.1
Jinja2==2.7
LinkedIn-Client-Library==1.0
M2Crypto==0.21.1
Mako==0.8.1
MarkupSafe==0.18
MySQL-python==1.2.4
PAM==0.4.2
PIL==1.1.7
Pillow==2.1.0
PyAudio==0.2.7
PyYAML==3.10
Pygments==1.6
Pyste==0.9.10
SOAPpy==0.12.5
SPARQLWrapper==1.5.2
SQLAlchemy==0.8.2
Skype4Py==1.0.35
UniConvertor==1.1.4
Unipath==1.0
adium-theme-ubuntu==0.3.3
alembic==0.6.0
apt-xapian-index==0.45
argparse==1.2.1
beautifulsoup4==4.2.1
blinker==1.3
boto==2.9.6
bzr==2.6b2
catkin-pkg==0.1.18
ccsm==0.9.9.0
cffi==0.6
chardet==2.1.1
compizconfig-python==0.9.9.0
configglue==1.1.0
configobj==4.7.2
configparser==3.3.0r2
cov-core==1.7
coverage==3.6
ctypes-configure==0.1
debtagshw==0.1
decorator==3.4.0
defer==1.0.6
dirspec==13.05
distribute==0.6.35
dnspython==1.10.0
docutils==0.10
duplicity==0.6.21
easy-thumbnails==1.4
empy==3.1
extras==0.0.3
fixtures==0.3.12
fpconst==0.7.2
gdata==2.0.17
guppy==0.1.10
html5lib==1.0b2
httplib2==0.8
httpretty==0.6.2
ipdb==0.7
ipython==0.13.2
isodate==0.4.9
itsdangerous==0.22
jedi==0.5b5
jockey==0.9.7
keyring==1.5
kitchen==1.1.1
launchpadlib==1.10.2
lazr.authentication==0.1.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
lxml==3.1.0
mechanize==0.2.5
mercurial==2.6.3
mock==1.0.1
msgpack-python==0.3.0
nltk==2.0.4
nose==1.3.0
numpy==1.7.1
oauth==1.0.1
oauth2==1.5.211
oauthlib==0.4.2
objgraph==1.7.2
oneconf==0.3.3
parallel==0.2.5
paramiko==1.10.1
parsedatetime==1.1.2
pep8==1.4.6
pexpect==2.4
phpsh==1.3.1
piston-mini-client==0.7.5
protobuf==2.5.0
py==1.4.15
pyOpenSSL==0.13
pyPdf==1.13
pybloomfiltermmap==0.3.11
pycountry==0.16
pycparser==2.09.1
pycrypto==2.6
pycups==1.9.62
pycurl==7.19.0
pyflakes==0.7.3
pygame==1.9.1release
pygeoip==0.2.6
pygobject==3.8.0
pygpgme==0.3
pyinotify==0.9.4
pyjulius==0.3
pymongo==2.5.2
pyparallel==0.2
pyparsing==1.5.7
pyserial==2.6
pysmbc==1.0.13
pysqlite==2.6.3
pytest==2.3.5
pytest-capturelog==0.7
pytest-codecheckers==0.2
pytest-cov==1.6
python-apt==0.8.8ubuntu6
python-augeas==0.4.1
python-crontab==1.4.1
python-dateutil==2.1
python-debian==0.1.21-nmu2ubuntu1
python-linkedin==4.0
python-memcached==1.53
python-mimeparse==0.1.4
python-nmap==0.2.7
python-openid==2.2.5
python-twitter==1.0
pyxdg==0.25
pyzmq==13.1.0
qtile==0.5
rdflib==4.0.1
redis==2.7.6
reportlab==2.7
requests==1.2.0
requests-oauthlib==0.3.3
rhythmbox-ubuntuone==4.2.0
scapy==2.2.0
screen-resolution-extra==0.0.0
selenium==2.31.0
sessioninstaller==0.0.0
simplegeneric==0.8.1
simplejson==3.3.0
six==1.3.0
sleekxmpp==1.1.11
software-center-aptd-plugins==0.0.0
sqlautocode==0.7
sure==1.2.2
system-service==0.1.6
testresources==0.2.7
testtools==0.9.32
tornado==3.1
twilio==3.5.1
twill==0.9
ubuntuone-storage-protocol==4.2.0
unittest2==0.5.1
unity-lens-photos==0.9
vboxapi==1.0
vcstools==0.1.31
virtkey==0.63.0
virtualenv==1.9.1
wadllib==1.3.2
wsgi-intercept==0.5.1
wsgiref==0.1.2
wstools==0.4.3
xdiagnose==3.5.1
xkit==0.0.0
zeroinstall-injector==2.0
zmqpy==0.3.2
zope.interface==4.0.5
